refactor: move entity file to repository
