--- example/spring-data-mongodb/src/main/kotlin/com/github/inflab/example/spring/data/mongodb/entity/sample/FruitQuantities.kt
@@ -1,7 +1,0 @@
-package com.github.inflab.example.spring.data.mongodb.entity.sample
-
-data class FruitQuantities(
-    val lemons: Int,
-    val oranges: Int,
-    val grapefruit: Int,
-)
