--- example/spring-data-mongodb/src/main/kotlin/com/github/inflab/example/spring/data/mongodb/entity/sample/Fruits.kt
@@ -1,13 +1,0 @@
-package com.github.inflab.example.spring.data.mongodb.entity.sample
-
-import org.springframework.data.annotation.Id
-import org.springframework.data.mongodb.core.mapping.Document
-
-@Document("fruits")
-data class Fruits(
-    @Id
-    val id: String,
-    val type: String?,
-    val description: String,
-    val quantities: FruitQuantities?,
-)
